psychat_ai/requirements.txt
llama-cpp-python
TTS
vosk
sounddevice
face_recognition
cryptography
opencv-python
ffmpeg-python
numpy
